+++ src/front/components/Footer.jsx
@@ -1,11 +1,9 @@
 export const Footer = () => (
-	<footer className="footer mt-5 py-3 text-center">
-		<p>
-			Check the <a target="_blank" href="https://4geeks.com/docs/start/react-flask-template">template documentation</a> <i className="fa-solid fa-file"></i> for help.
-		</p>
-		<p>
-			Made with <i className="fa fa-heart text-danger" /> by{" "}
-			<a href="http://www.4geeksacademy.com">4Geeks Academy</a>
-		</p>
+	<footer className="bg-light border-top py-3 text-center">
+		<div className="container">
+			<p className="mb-0">
+				<small>&copy; {new Date().getFullYear()} αlpha. Todos los derechos reservados.</small>
+			</p>
+		</div>
 	</footer>
 );
