+++ src/front/index.css
@@ -1,3 +1,9 @@
+#root {
+  display: grid;
+  min-height: 100dvh;
+  grid-template-rows: auto 1fr auto;
+}
+
 .unsetBtn {
   all: unset;
 }
